All tests passing on MRI 2.0
Added support for Binding#frame_type, via frame_iseq in RubyVM::DebugInspector.
Note it just appeared like the 10th element in iseq.to_a returns the frame type
but this could potentially be unreliable, though it seems to work fine. :D

+++ Rakefile
@@ -3,20 +3,23 @@ direc = File.dirname(__FILE__)
 
 $:.unshift 'lib'
 
-PROJECT_NAME = "binding_of_caller"
-
 require 'rake/clean'
 require 'rubygems/package_task'
 
-require "#{PROJECT_NAME}/version"
+require "binding_of_caller/version"
 
 CLOBBER.include("**/*.#{dlext}", "**/*~", "**/*#*", "**/*.log", "**/*.o")
 CLEAN.include("ext/**/*.#{dlext}", "ext/**/*.log", "ext/**/*.o",
               "ext/**/*~", "ext/**/*#*", "ext/**/*.obj", "**/*#*", "**/*#*.*",
               "ext/**/*.def", "ext/**/*.pdb", "**/*_flymake*.*", "**/*_flymake", "**/*.rbc")
 
+def mri_2?
+  defined?(RUBY_ENGINE) && RUBY_ENGINE == "ruby" &&
+    RUBY_VERSION =~ /^2/
+end
+
 def apply_spec_defaults(s)
-  s.name = PROJECT_NAME
+  s.name = "binding_of_caller"
   s.summary = "Retrieve the binding of a method's caller. Can also retrieve bindings even further up the stack."
   s.version = BindingOfCaller::VERSION
   s.date = Time.now.strftime '%Y-%m-%d'
@@ -24,8 +27,9 @@ def apply_spec_defaults(s)
   s.email = 'jrmair@gmail.com'
   s.description = s.summary
   s.require_path = 'lib'
-  s.add_development_dependency('bacon')
-  s.add_development_dependency('rake')
+  s.add_dependency 'debug_inspector', '~> 0.0.1'
+  s.add_development_dependency 'bacon'
+  s.add_development_dependency 'rake'
   s.homepage = "http://github.com/banister/binding_of_caller"
   s.has_rdoc = 'yard'
   s.files = `git ls-files`.split("\n")
@@ -67,7 +71,7 @@ namespace :ruby do
   spec = Gem::Specification.new do |s|
     apply_spec_defaults(s)
     s.platform = Gem::Platform::RUBY
-    s.extensions = ["ext/#{PROJECT_NAME}/extconf.rb"]
+    s.extensions = ["ext/binding_of_caller/extconf.rb"]
   end
 
   Gem::PackageTask.new(spec) do |pkg|
@@ -84,27 +88,30 @@ namespace :ruby do
 end
 
 desc "build the binaries"
-task :compile do
-  chdir "./ext/#{PROJECT_NAME}/" do
-    sh "ruby extconf.rb"
-    sh "make clean"
-    sh "make"
-    sh "cp *.#{dlext} ../../lib/"
+task :compile => :cleanup do
+  if !mri_2?  
+    chdir "./ext/binding_of_caller/" do
+      sh "ruby extconf.rb"
+      sh "make"
+      sh "cp *.#{dlext} ../../lib/"
+    end
   end
 end
 
 desc 'cleanup the extensions'
 task :cleanup do
-  sh 'rm -rf lib/binding_of_caller.so'
-  chdir "./ext/#{PROJECT_NAME}/" do
-    sh 'make clean'
+  if !mri_2?
+    sh 'rm -rf lib/binding_of_caller.so'
+    chdir "./ext/binding_of_caller/" do
+      sh 'make clean' 
+    end
   end
 end
 
 desc "reinstall gem"
 task :reinstall => :gems do
   sh "gem uninstall binding_of_caller" rescue nil
-  sh "gem install #{direc}/pkg/#{PROJECT_NAME}-#{BindingOfCaller::VERSION}.gem"
+  sh "gem install #{direc}/pkg/binding_of_caller-#{BindingOfCaller::VERSION}.gem"
 end
 
 task :install => :reinstall
